requirements.txt
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==5.0.0
boto3==1.42.16
botocore==1.42.16
cachetools==6.2.4
cairocffi==1.7.1
CairoSVG==2.8.2
Cartopy==0.25.0
certifi==2025.11.12
cffi==2.0.0
cfgrib==0.9.15.1
cftime==1.6.5
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
contourpy==1.3.2
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
Django==3.2.25
django-environ==0.12.0
django-model-utils==5.0.0
djangorestframework==3.15.1
dotenv==0.9.9
eccodes==2.44.0
eccodeslib==2.44.1.8
ecdsa==0.19.1
eckitlib==1.32.4.8
exceptiongroup==1.3.1
fastapi==0.124.0
fckitlib==0.14.1.8
feedgen==1.0.0
findlibs==0.1.2
fonttools==4.61.0
frozenlist==1.8.0
geojson==3.2.0
geopandas==1.1.1
google-auth==2.45.0
google-genai==1.56.0
h11==0.16.0
herbie==1.0b2
herbie-data==2025.12.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.7
multidict==6.7.0
netCDF4==1.7.3
numpy==2.2.6
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
propcache==0.4.1
psutil==7.2.1
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.1
pyshp==3.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytz==2025.2
rasterio==1.4.4
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
starlette==0.50.0
strict-rfc3339==0.7
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.5.1
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.6.0
uvicorn==0.38.0
webencodings==0.5.1
websockets==15.0.1
xarray==2025.6.1
xgboost==3.1.2
yarl==1.22.0
